#217, #202 - Upgrade backbone-nested to 2.0.4, not stringifying idsToFetch on cache fetchees, and removed unneeded test console logs
