+++ modules/mixins/cacheMixin.js
@@ -355,7 +355,7 @@
               type: collection.fetchHttpAction,
               url: _.result(collection, 'url') + collection.getByIdsUrl,
               contentType: 'application/json; charset=utf-8',
-              data: JSON.stringify(idsToFetch)
+              data: idsToFetch
             }).done(
               // Success function
               function(data) {
@@ -446,30 +446,6 @@
           });
         }
       },
-
-      /**
-       * Will abolish all listeners and events that are hooked
-       * to this collection.
-       * @method dispose
-       */
-      dispose: function() {
-        this.unbind();
-        this.off();
-        this.stopListening();
-        this.stopPolling();
-        if (this.isRequester) {
-          this.requesterDispose();
-        }
-      },
-
-      /**
-       * The default filter.  Always returns itself.
-       * @method filterDefault
-       * @return {Collection} a new instance of this collection
-       */
-      filterDefault: function() {
-        return this.constructor(this);
-      }
     };
   };
