+++ package.json
@@ -19,7 +19,7 @@
     "jquery": "2.1.3",
     "underscore": "1.8.2",
     "backbone": "1.2.3",
-    "backbone-nested": "2.0.3",
+    "backbone-nested": "2.0.4",
     "backbone.stickit": "0.9.2"
   },
   "devDependencies": {
